lock_server/requirements.txt
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
gunicorn==20.0.0
idna==2.8
image==1.5.27
Pillow==9.3.0
pycryptodome==3.9.2
PyMySQL==0.9.3
pyotp==2.3.0
pytz==2019.3
qrcode==6.1
requests==2.22.0
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
supervisor==4.1.0
urllib3==1.26.5
